# Conflicts:
#	run_data_generator.py

+++ CartPole/state_utilities.py
@@ -45,7 +45,7 @@ def create_cartpole_state(state: dict = {}, dtype=None) -> np.ndarray:
     if dtype is None:
         dtype = np.float32
 
-    s = np.zeros_like(STATE_VARIABLES, dtype=dtype)
+    s = np.zeros_like(STATE_VARIABLES, dtype=np.float32)
     for i, v in enumerate(STATE_VARIABLES):
         s[i] = state.get(v) if v in state.keys() else s[i]
     return s
